[BE][Fix] s3 이미지 null인 경우 처리 방법 수정

See merge request s10-final/S10P31D104!129

+++ backend/src/main/java/com/todayeat/backend/_common/util/S3Util.java
@@ -29,7 +29,11 @@ public class S3Util {
     private String bucket;
 
     // S3 버킷에 이미지 파일 업로
-    public String uploadImage(MultipartFile multipartFile, DirectoryType directoryType, Long dirNamePrincipalId) {
+    public String uploadImageIfPresent(MultipartFile multipartFile, DirectoryType directoryType, Long dirNamePrincipalId) {
+
+        if (multipartFile == null) {
+            return null;
+        }
 
         // 이미지 형식의 파일인지 확인
         if (!Objects.requireNonNull(multipartFile.getContentType()).contains("image")) {
@@ -53,22 +57,21 @@ public class S3Util {
     }
 
     //S3 버킷의 이미지 객체 삭제, 이미지 수정시 기존 이미지 제거에 사용
-    public void deleteImage(String fileUrl) {
+    public void deleteImageIfPresent(String fileUrl) {
+
+        // fileUrl이 null이면 삭제 로직 안한다.
+        if (fileUrl == null) {
+            log.error("fileUrl이 null입니다.");
+            return;
+        }
 
         String key = getS3ObjetKey(fileUrl);
+
         try {
             amazonS3.deleteObject(bucket, key);
         } catch (AmazonServiceException e) {
             log.error("S3Util deleteImage deleteObject error : {}, {}", e, key);
             throw new BusinessException(ErrorType.INTERNAL_SERVER_ERROR);
-        }
-    }
-
-    public Optional<String> getOptionalS3ObjetKey(String fileUrl) {
-        try {
-            return Optional.of(getS3ObjetKey(fileUrl));
-        } catch (Exception e) {
-            return Optional.empty();
         }
     }
 
@@ -101,7 +104,8 @@ public class S3Util {
         int startIndex = fileUrl.indexOf('/', fileUrl.indexOf("//") + 2);
 
         if (startIndex < 0) {
-            throw new BusinessException(ErrorType.IMAGE_URL_FORMAT_INVALID);
+            log.error("s3 key 형식이 아닙니다.");
+            return null;
         }
 
         return fileUrl.substring(startIndex + 1);
